Revert "Addressing PR comments: parallelizing node processing, adding logging for unsupported types, and moving tests to a new file"
This reverts commit 917e96f52144e24af60dcbae8d690ee2ed286b8c.
